+++ src/helpers/router.js
@@ -1,19 +1,21 @@
-import {Observable} from 'rx'
+// now in 'util'
 
-function nestedComponent({path$, value$}, sources) {
-  return path$.zip(value$,
-    (path, value) => value({...sources, router: sources.router.path(path)})
-  ).shareReplay(1)
-}
+// import {Observable} from 'rx'
 
-const mergeOrFlatMapLatest = (prop, ...sourceArray) =>
-  Observable.merge(
-    sourceArray.map(src => // array map not observable!
-      typeof src.subscribe === 'function' ?
-        src.flatMapLatest(l => l[prop] || Observable.empty()) :
-        // otherwise look for a prop
-        src[prop] || Observable.empty()
-    )
-  )
+// function nestedComponent({path$, value$}, sources) {
+//   return path$.zip(value$,
+//     (path, value) => value({...sources, router: sources.router.path(path)})
+//   ).shareReplay(1)
+// }
 
-export {nestedComponent, mergeOrFlatMapLatest}
+// const mergeOrFlatMapLatest = (prop, ...sourceArray) =>
+//   Observable.merge(
+//     sourceArray.map(src => // array map not observable!
+//       typeof src.subscribe === 'function' ?
+//         src.flatMapLatest(l => l[prop] || Observable.empty()) :
+//         // otherwise look for a prop
+//         src[prop] || Observable.empty()
+//     )
+//   )
+
+// export {nestedComponent, mergeOrFlatMapLatest}
